Fix: final fix

+++ src/main/java/com/hng/googleAuth/models/PaymentRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class PaymentRequest {
-    private String email; // customer email (required by Paystack)
-    private Long amount; // amount in kobo
+    private String email;
+    private Double amount;
 }
